+++ modules/i3.nix
@@ -60,7 +60,7 @@
         "Mod4+l" = "--release exec ${pkgs.xorg.xset}/bin/xset dpms force off";
 
         "${mod}+Shift+Return" =
-          "exec ${pkgs.alacritty}/bin/alacritty -o window.class.instance=floating";
+          "exec ${pkgs.alacritty}/bin/alacritty -o 'window.class.instance=\"floating\"'";
 
         # switch between workspaces on the current monitor
         "${mod}+1" = "exec ${i3-ws} --ws 1";